hardware-gateway/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
redis==4.6.0
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0